Time: 5 ms (67.37%), Space: 38.7 MB (50.59%) - LeetHub

+++ 1708-design-parking-system/1708-design-parking-system.cpp
@@ -1,6 +1,6 @@
 class ParkingSystem {
 private:
-    short cars[4];
+    int cars[4];
 public:
     ParkingSystem(int big, int medium, int small) {
         cars[1] = big;
